+++ ChangeLog.md
@@ -5,7 +5,8 @@ The full changelog of the platform includes also changes brought to the Semantic
 https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 10.2.2 (dd-mm-yyyy)
-  * 
+  * Added SuperUser system role
+  * Allowed SuperUser to create projects where it is automatically set as Project Manager
 
 # 10.2.1 (02-07-2022)
   * Added checkbox for ignoring semantic checks in Manchester editor